Hg-commit: 2a8e32b53b253a4b939d0602c9e0c2755d109f2b

+++ src/sys/plog/plog.c
@@ -1920,7 +1920,7 @@ PetscErrorCode PETSCSYS_DLLEXPORT PetscGetFlops(PetscLogDouble *flops)
 PetscErrorCode PETSCSYS_DLLEXPORT PetscLogObjectState(PetscObject obj, const char format[], ...)
 {
   PetscErrorCode ierr;
-  int            fullLength;
+  size_t         fullLength;
   va_list        Argp;
 
   PetscFunctionBegin;
